+++ src/Core/Responses/Response.cpp
@@ -1,23 +1,3 @@
 #include "../../../inc/Core/Responses/Response.h"
 
 template <typename ResType> Response<ResType>::~Response() {}
-
-DBResponse::DBResponse(const pqxx::result &ResponseData)
-    : m_ResponseData{ResponseData}, m_ResponseSize{sizeof(ResponseData)} {
-  APP_INFO("Response Size -> " + std::to_string(m_ResponseSize));
-}
-
-DBResponse::DBResponse(pqxx::result &&ResponseData)
-    : m_ResponseData{std::move(ResponseData)},
-      m_ResponseSize{sizeof(ResponseData)} {
-  APP_INFO("Response Size -> " + std::to_string(m_ResponseSize));
-}
-
-DBResponse::DoubleDuration
-DBResponse::RunBenchmark(std::function<void()> Func) {
-  auto start = Clock::now();
-  Func();
-  auto end = Clock::now();
-  auto duration = end - start;
-  return duration;
-}
